refactor(stubs): split _html.pyi into a _stubs subsystem package
Break the 460-line flat extension stub into _stubs/{tokenizer,dom,serialize,
query,features}.pyi mirroring the _c/ tree, with _html.pyi kept as a single
file that re-exports them, so it stays the stub for _html.so and the API map
is navigable. ty reports no new diagnostics; runtime and tests unchanged.
