+++ src/turbohtml/_stubs/__init__.pyi
@@ -1,0 +1,1 @@
+# Type-stub package mirroring the _c/ subsystems of the compiled turbohtml._html extension.
