+++ src/turbohtml/_stubs/dom.pyi
@@ -1,0 +1,348 @@
+# Subsystem: dom (_c/dom) — the node model, parsing entrypoints, and tree reconstruction.
+from collections.abc import Callable, Iterable, Iterator, Mapping, MutableMapping, Sequence
+from enum import Enum, IntEnum
+from re import Pattern
+from typing import Literal, TypeAlias, final
+
+from .._links import Link
+from .serialize import Formatter, Indent, Minify
+
+_Filter: TypeAlias = str | Pattern[str] | bool | Callable[[str | None], bool] | list[_Filter]
+
+class Axis(IntEnum):
+    DESCENDANTS = 0
+    CHILDREN = 1
+    ANCESTORS = 2
+    NEXT_SIBLINGS = 3
+    PREVIOUS_SIBLINGS = 4
+    FOLLOWING = 5
+    PRECEDING = 6
+
+class Namespace(Enum):
+    HTML = "html"
+    SVG = "svg"
+    MATHML = "math"
+
+class Node:
+    @property
+    def parent(self) -> Element | Document | None: ...
+    @property
+    def children(self) -> Sequence[Node]: ...
+    @property
+    def next_sibling(self) -> Node | None: ...
+    @property
+    def previous_sibling(self) -> Node | None: ...
+    @property
+    def descendants(self) -> Iterator[Node]: ...
+    @property
+    def ancestors(self) -> Iterator[Element | Document]: ...
+    @property
+    def next_siblings(self) -> Iterator[Node]: ...
+    @property
+    def previous_siblings(self) -> Iterator[Node]: ...
+    @property
+    def following(self) -> Iterator[Node]: ...
+    @property
+    def preceding(self) -> Iterator[Node]: ...
+    @property
+    def strings(self) -> Iterator[str]: ...
+    @property
+    def stripped_strings(self) -> Iterator[str]: ...
+    @property
+    def text(self) -> str: ...
+    @property
+    def html(self) -> str: ...
+    @property
+    def inner_html(self) -> str: ...
+    @property
+    def source_line(self) -> int | None: ...
+    @property
+    def source_col(self) -> int | None: ...
+    @property
+    def position(self) -> tuple[int, int] | None: ...
+    def find(
+        self,
+        tag: _Filter | None = None,
+        /,
+        *,
+        axis: Axis = ...,
+        attrs: Mapping[str, _Filter] | None = ...,
+        class_: _Filter | None = ...,
+        **filters: _Filter,
+    ) -> Element | None: ...
+    def find_all(
+        self,
+        tag: _Filter | None = None,
+        /,
+        *,
+        axis: Axis = ...,
+        attrs: Mapping[str, _Filter] | None = ...,
+        class_: _Filter | None = ...,
+        limit: int | None = ...,
+        **filters: _Filter,
+    ) -> list[Element]: ...
+    def select(self, selector: str, /) -> list[Element]: ...
+    def select_one(self, selector: str, /) -> Element | None: ...
+    def xpath(
+        self,
+        expression: str,
+        /,
+        *,
+        smart_strings: bool = ...,
+        extensions: dict[tuple[str | None, str], Callable[..., str | float | bool]] | None = ...,
+        **variables: str | float | bool,
+    ) -> list[Element | str]: ...
+    def xpath_iter(
+        self,
+        expression: str,
+        /,
+        *,
+        smart_strings: bool = ...,
+        extensions: dict[tuple[str | None, str], Callable[..., str | float | bool]] | None = ...,
+        **variables: str | float | bool,
+    ) -> Iterator[Element | str]: ...
+    def xpath_one(
+        self,
+        expression: str,
+        /,
+        *,
+        smart_strings: bool = ...,
+        extensions: dict[tuple[str | None, str], Callable[..., str | float | bool]] | None = ...,
+        **variables: str | float | bool,
+    ) -> Element | str | None: ...
+    def matches(self, selector: str, /) -> bool: ...
+    def closest(self, selector: str, /) -> Element | None: ...
+    def prune(self, selector: str, /) -> Node: ...
+    def re(self, pattern: str | Pattern[str], /, *, attr: str | None = None) -> list[str]: ...
+    def re_first(
+        self, pattern: str | Pattern[str], /, default: str | None = None, *, attr: str | None = None
+    ) -> str | None: ...
+    def serialize(
+        self,
+        *,
+        formatter: Formatter = ...,
+        layout: Indent | Minify | None = ...,
+        sort_attributes: bool = ...,
+        meta_charset: bool = ...,
+    ) -> str: ...
+    def encode(
+        self,
+        encoding: str = "utf-8",
+        *,
+        formatter: Formatter = ...,
+        layout: Indent | Minify | None = ...,
+        sort_attributes: bool = ...,
+        meta_charset: bool = ...,
+    ) -> bytes: ...
+    def to_markdown(
+        self,
+        *,
+        heading_style: Literal["atx", "atx_closed", "setext"] = "atx",
+        bullets: str = "-",
+        strong: str = "**",
+        emphasis: str = "*",
+        strikethrough: Literal["keep", "hide"] = "keep",
+        ignore_emphasis: bool = False,
+        sub_symbol: str = "",
+        sup_symbol: str = "",
+        code_block_style: Literal["fenced", "indented"] = "fenced",
+        code_language: str = "",
+        mark_code: bool = False,
+        link_style: Literal["inline", "reference"] = "inline",
+        autolink: bool = True,
+        link_title: bool = False,
+        ignore_links: bool = False,
+        skip_internal_links: bool = False,
+        base_url: str = "",
+        image_mode: Literal["markdown", "alt", "ignore", "html"] = "markdown",
+        default_image_alt: str = "",
+        table_mode: Literal["markdown", "strip", "html"] = "markdown",
+        table_header: Literal["first", "detect", "none"] = "first",
+        pad_tables: bool = False,
+        escape_mode: Literal["minimal", "all"] = "minimal",
+        escape_asterisks: bool = True,
+        escape_underscores: bool = True,
+        line_break: Literal["spaces", "backslash"] = "spaces",
+        block_spacing: Literal["double", "single"] = "double",
+        wrap_width: int = 0,
+        wrap_list_items: bool = False,
+        wrap_links: bool = True,
+        transliterate: bool = False,
+        document_strip: Literal["strip", "lstrip", "rstrip", "none"] = "strip",
+        quote_open: str = '"',
+        quote_close: str = '"',
+        google_doc: bool = False,
+        google_list_indent: int = 36,
+        hide_strikethrough: bool = False,
+        strip: Iterable[str] | None = None,
+        convert: Iterable[str] | None = None,
+        converters: Mapping[str, Callable[[Element, str], str]] | None = None,
+    ) -> str: ...
+    def to_text(
+        self,
+        *,
+        width: int = 0,
+        links: Literal["none", "inline", "footnote"] = "none",
+        images: bool = False,
+        layout: Literal["extended", "strict"] = "extended",
+        default_image_alt: str = "",
+        table_cell_separator: str = "  ",
+        bullet: str = "* ",
+    ) -> str: ...
+    def to_annotated_text(
+        self,
+        annotation_rules: Mapping[str, Sequence[str]],
+        /,
+        *,
+        width: int = 0,
+        links: Literal["none", "inline", "footnote"] = "none",
+        images: bool = False,
+        layout: Literal["extended", "strict"] = "extended",
+        default_image_alt: str = "",
+        table_cell_separator: str = "  ",
+        bullet: str = "* ",
+    ) -> tuple[str, list[tuple[int, int, str]]]: ...
+    def links(self) -> list[Link]: ...
+    def rewrite_links(self, replace: Callable[[str], str | None], /) -> None: ...
+    def resolve_links(self, base_url: str, /) -> None: ...
+    def main_content(self) -> Element | None: ...
+    def main_text(self) -> str: ...
+    def insert_before(self, *nodes: Node) -> None: ...
+    def insert_after(self, *nodes: Node) -> None: ...
+    def replace_with(self, *nodes: Node) -> None: ...
+    def wrap(self, wrapper: Element, /) -> Element: ...
+    def wrap_siblings(self, wrapper: Element, /, *, until: Node | None = None) -> Element: ...
+    def unwrap(self) -> Node: ...
+    def extract(self) -> Node: ...
+    def decompose(self) -> None: ...
+    def __iter__(self) -> Iterator[Node]: ...
+    def __len__(self) -> int: ...
+    def __getitem__(self, index: int) -> Node: ...
+    def __bool__(self) -> bool: ...
+    def __hash__(self) -> int: ...
+    def __eq__(self, other: object) -> bool: ...
+
+@final
+class Element(Node):
+    __match_args__ = ("tag",)
+    def __init__(self, tag: str, attrs: Mapping[str, str | list[str] | None] | None = None) -> None: ...
+    @property
+    def tag(self) -> str: ...
+    @property
+    def namespace(self) -> Namespace: ...
+    @property
+    def attrs(self) -> MutableMapping[str, str | list[str] | None]: ...
+    @property
+    def text(self) -> str: ...
+    @text.setter
+    def text(self, value: str) -> None: ...
+    @property
+    def field_value(self) -> str | list[str] | None: ...
+    @field_value.setter
+    def field_value(self, value: str | list[str] | None) -> None: ...
+    @property
+    def checked(self) -> bool: ...
+    @checked.setter
+    def checked(self, value: bool) -> None: ...
+    def form_data(self) -> list[tuple[str, str]]: ...
+    def attr(self, name: str, /, default: str | None = None) -> str | None: ...
+    def append(self, child: Node, /) -> None: ...
+    def extend(self, children: Iterable[Node], /) -> None: ...
+    def insert(self, index: int, child: Node, /) -> None: ...
+    def clear(self) -> None: ...
+    def normalize(self) -> None: ...
+    def wrap_children(self, wrapper: Element, /) -> Element: ...
+
+@final
+class Text(Node):
+    __match_args__ = ("data",)
+    def __init__(self, data: str) -> None: ...
+    @property
+    def data(self) -> str: ...
+    @data.setter
+    def data(self, value: str) -> None: ...
+
+@final
+class Comment(Node):
+    __match_args__ = ("data",)
+    def __init__(self, data: str) -> None: ...
+    @property
+    def data(self) -> str: ...
+    @data.setter
+    def data(self, value: str) -> None: ...
+
+@final
+class CData(Node):
+    __match_args__ = ("data",)
+    def __init__(self, data: str) -> None: ...
+    @property
+    def data(self) -> str: ...
+    @data.setter
+    def data(self, value: str) -> None: ...
+
+@final
+class ProcessingInstruction(Node):
+    __match_args__ = ("target", "data")
+    def __init__(self, target: str, data: str) -> None: ...
+    @property
+    def target(self) -> str: ...
+    @property
+    def data(self) -> str: ...
+
+@final
+class Doctype(Node):
+    __match_args__ = ("name",)
+    @property
+    def name(self) -> str: ...
+    @property
+    def public_id(self) -> str | None: ...
+    @property
+    def system_id(self) -> str | None: ...
+
+@final
+class ParseError:
+    @property
+    def code(self) -> str: ...
+    @property
+    def line(self) -> int: ...
+    @property
+    def col(self) -> int: ...
+    def __hash__(self) -> int: ...
+    def __eq__(self, other: object) -> bool: ...
+
+class HTMLParseError(Exception):
+    error: ParseError
+
+@final
+class Document(Node):
+    __match_args__ = ("root",)
+    @property
+    def root(self) -> Element | None: ...
+    @property
+    def encoding(self) -> str | None: ...
+    def base_url(self, fallback: str = "") -> str: ...
+    def meta_refresh(self, fallback: str = "") -> tuple[float, str] | None: ...
+    @property
+    def errors(self) -> list[ParseError]: ...
+
+@final
+class IncrementalParser:
+    def __init__(self, *, encoding: str = "utf-8", positions: bool = True) -> None: ...
+    def feed(self, data: str | bytes) -> None: ...
+    def close(self) -> Document: ...
+    def __enter__(self) -> IncrementalParser: ...
+    def __exit__(self, *exc: object) -> None: ...
+
+def parse(
+    markup: str | bytes,
+    *,
+    encoding: str | None = None,
+    strict: bool = False,
+    detect_encoding: bool = False,
+    positions: bool = True,
+) -> Document: ...
+def parse_fragment(html: str, context: str = "div", *, positions: bool = True) -> Element: ...
+def _parse_tree(html: str, /) -> str: ...
+def _parse_fragment(html: str, context: str, /) -> str: ...
+def _parse_only(html: str, /) -> None: ...
+def _reconstruct(kind: int, descriptor: tuple[object, ...], children: list[Node], /) -> Node: ...
