+++ src/turbohtml/_stubs/query.pyi
@@ -1,0 +1,3 @@
+# Subsystem: query (_c/query) — XPath compilation hooks and smart-string registration.
+def _xpath_parse(expression: str, /) -> str: ...
+def _register_xpath_string(xpath_string_type: type, /) -> None: ...
