+++ src/turbohtml/_stubs/tokenizer.pyi
@@ -1,0 +1,55 @@
+# Subsystem: tokenizer (_c/tokenizer) — the HTML tokenization surface.
+from collections.abc import Iterator
+from enum import IntEnum
+from typing import final
+
+class TokenType(IntEnum):
+    TEXT = 0
+    START_TAG = 1
+    END_TAG = 2
+    COMMENT = 3
+    DOCTYPE = 4
+    CHARACTER_REFERENCE = 5
+
+@final
+class Token:
+    @property
+    def type(self) -> TokenType: ...
+    @property
+    def data(self) -> str | None: ...
+    @property
+    def source(self) -> str | None: ...
+    @property
+    def tag(self) -> str | None: ...
+    @property
+    def attrs(self) -> list[tuple[str, str]] | None: ...
+    @property
+    def self_closing(self) -> bool: ...
+    @property
+    def name(self) -> str | None: ...
+    @property
+    def public_id(self) -> str | None: ...
+    @property
+    def system_id(self) -> str | None: ...
+    @property
+    def force_quirks(self) -> bool: ...
+    @property
+    def line(self) -> int: ...
+    @property
+    def col(self) -> int: ...
+    def attr(self, name: str, default: str | None = None) -> str | None: ...
+
+@final
+class Tokenizer:
+    def __init__(self, *, resolve_references: bool = ..., capture_source: bool = ...) -> None: ...
+    def feed(self, data: str) -> Iterator[Token]: ...
+    def close(self) -> Iterator[Token]: ...
+    def reset(self) -> None: ...
+    def __enter__(self) -> Tokenizer: ...
+    def __exit__(self, *exc: object) -> None: ...
+    def __iter__(self) -> Iterator[Token]: ...
+
+def tokenize(s: str, /, *, resolve_references: bool = ..., capture_source: bool = ...) -> Iterator[Token]: ...
+def _tokenize_states(
+    text: str, initial_state: str, last_start_tag: str | None = ..., storage_kind: int = ..., /
+) -> list[tuple[object, ...]]: ...
